Remove the need for `no_force`
